Reduce London from 30 to 27 cells
We maintain enough cells to satisfy our tenant demand for app memory
and our reliability goals [1]. We have not needed more than 27 cells
in London since June 24th, and most of the time we would be fine with
just 24.

This commit reduces us to 27 cells. We expect demand in London to
gradually climb, so it makes sense to allow some headroom over that
24-ish requirement.

[1] https://team-manual.cloud.service.gov.uk/architecture_decision_records/ADR021-cell-capacity-assignment-2/