Explain our OAuth configurations for posterity
This commit adds some security explanation to the username config of
our OAuth Identity Provider configurations. These are used to allow
sign-in to the PaaS with Google or Microsoft.

Without these comments there is no obvious record that UAA allows
login based upon the `email` claim. That knowledge is critical when
considering allowing a new OAuth provider, or assessing the security
of how we configure the current ones.