Metrics for AWS costs (values delayed by 1 day)
This commit adds new metrics to our `paas-metrics` app. These metrics
go into Prometheus. These new metrics are to monitor our spend in AWS:

1. `aws.cost_explorer.by_service` gives the cost of each AWS service
   in the AWS region that the current PaaS environment is using.

2. `aws.cost_explorer.by_region` gives the total cost of every AWS
   region.

The values for this metric are for the entire previous day. The way
that a Prometheus library is used does not support providing
timestamps, which is annoying. Also annoying is that AWS Cost Explorer
is only updated "a few times a day" so we can't provide real-time data
and instead are reporting for all of the previous day.

1-day-delayed costs of each AWS service
   in the AWS region

2.

how much we spend in every region of the current AWS account, and how

The purpose of these metrics is to report our AWS spend so that we can
keep an eye on it and try to alert if it differs drastically from what
we are rebilling to tenants.

The metrics are filed for the current time but two days in the past,
and are accompanied with the full cost of that day two days ago. This
means we're reporting two days in the past. AWS's documentation only
says that Cost Explorer updates a few times a day, so adding a bit of
a delay will help stabilise things.