Add `cf-shell` script to scripts
cf-shell logs in to a CF deployment using the cf_subshell_scoped_login and
cf_admin_login_dev scripts as appropriate (based on environment name).
Optionally allows the user to force the login to happen inside a subshell for
all environments.

This is useful for preventing us from being accidentally still logged in to an
environment when issuing `cf` commands.

The scripts expose a $CF_SUBSHELL_TARGET which holds the name of the
environment being targetted. This can be used in developers terminals to show
them which environment they're in.