Add USD currency metrics to paas-metrics
Unforeseen Political Events have caused the value of our currency (Great
British Pounds) to change not insignificantly.

The exchange rate is not something we keep in our heads regularly, but
has a significant impact on cross-charging and re-charging for
infrastructure; a large portion of our infrastructure costs are in USD,
so we risk losing money.

Add 2 metrics, one which is hardcoded to the same value as our exchange
rate, the other uses the European Central Bank currency API to get a
GBP/USD exchange rate.

We can, in a later commit, add an alert which checks that these have not
diverged significantly.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>