Convert billing logs from elasticsearch to metrics
There are three things that paas-billing's collector does that sometimes
take a potentially problematic length of time:

* Run the "create_events.sql" script
* Run the "create_billable_event_components.sql" script
* Consolidate the past month's events

We can see how long these things are taking in the logs, but we only
have a 30 day retention on those. If we're going to spot longer term
trends on paas-billing's performance we'll need to persist these metrics
for longer, so it makes sense to put them in prometheus.

This commit uses the Elasticsearch endpoint provided by Logit.io to
search for recent performance logs, averages them over the time window,
and reports them as metrics.

The three slow things that we're interested in happen on different
cadences, so we're looking back over different windows for different
things.

* The "create_events.sql" script runs every 30 minutes
* The "create_billable_event_components.sql" script runs every 30 minutes
* The monthly consolidation runs every month

To avoid having gap-y metrics we're querying Elasticsearch for a longer
time period so we're less likely to miss a result.