Switch from UnblendedCost to AmortizedCost
Because we reserve instances, the UnblendedCost appears artificially low
(since most of our instances were paid for back in July).

This makes comparing the money we're spending on AWS with the money
we're getting from billing very difficult - it would be much easier if
we reported the AmortizedCost metric instead.