Reduce thresholds for RDS billing alerts
These are currently alerting because we're only billing for ~70% of our
RDS usage, but this may actually be reasonable because we run some RDS
instances ourselves.

In any case we don't think billing is broken right now, so it seems
reasonable to reduce the thresholds so the alerts will only fire if it
actually breaks.

I've also made the upper bound lower because there should never be a
situation where we're billing more than the cost of RDS.