Update exchange rate to match current average
It's been 0.8 since at least March 2018
(alphagov/paas-billing@838212be12318a4c03598666e525fb5b9ea899b1). For
most of that time the actual rate has been quite a bit lower than that,
but now it's more like 0.82.

Updating it to match so we don't undercharge people for the AWS
resources we're using.

This will affect all non-consolidated bills (so everything from the
month this is merged onwards).