Upload secrets to credhub instead of S3
We're going to start using credhub for handling these secrets instead of
S3. As of this commit the pipelines will still be using the secrets from
S3 - we'll update the pipelines to use credhub in subsequent commits.

The thing that's not very nice about this is having to use docker to do
the actual upload. This is a bit tricky to do straight from ruby because
there's some relatively complicated networking going on (basically you
need an SSH tunnel to a SOCKS5 proxy to be open to be able to connect to
the credhub API). We've switched from bash to ruby otherwise to make the
scripts a bit more maintainable.