Upgrade CF RDS instance class and storage
AWS give databases IOPS in proportion to their allocated storage. They
don't recommend giving DBs less than 100G due to IOPs restrictions (10G
implies a limit of 30 operations / second, which is not great).

We're also seeing our London database hitting its CPU limits, so it's
worth upgrading the instance size too. May as well move to the latest
instance class (m5) while we're at it.

Price was 268.02 USD, will be 542.76 USD (twice for the production
environments, so an extra $548 / month total).

We're upgrading dev and staging instance types to m5, but leaving them
on the large size (as opposed to xlarge for production) as these are
sufficient at the moment, and we don't want to pay for more than we need.

We are upgrading the storage everywhere because it's not very expensive,
but working out why your DB has run out of IOPS is.