Correct TLS version acceptance tests for ALBs
When connecting to a classic load balancer with an unsupported TLS version,
Go's TLS library can report that the version is unsupported. However, when
connecting to an ALB with an unsupported TLS version, the ALB simply closes the
connection.

This commit allows for that difference in the custom acceptance tests. Checking
for a connection reset message isn't a perfect test, because it can show up as
a false positive, but if the ALB is just closing the connection then this is
the best we can do.