Continously deploy paas-auditor
Some time ago, commit cca2ae3252fa520f659b9b0d486bff048a17cbd3 added a
UAA client for Miki's paas-auditor project [1]. Miki deployed it
manually in production as getting it continuously-deployed isn't a
priority.

This commit implements that continuous deployment for paas-auditor.
To minimise the maintenance burden we don't pin to a particular
revision in `paas-cf`; instead we deploy `master` and require it to be
signed with a preapproved GPG key. This is similar to what is already
done for `paas-cf`, `paas-billing` and `paas-admin`.

[1] https://github.com/alphagov/paas-auditor