Read paas-pass secrets from credhub in the pipeline
Now that we're uploading the secrets to credhub instead of S3 we can use
concourse and bosh's config-servers to read the parameters using the
`((secret_name))` syntax.

This simplifies the pipeline a bit because we used to have to download
things from S3, set them as environment variables and pass them the
bosh. Now we can just let bosh look them up from credhub :tada:.