Relax billing alerts so they don't constantly fire
These are currently firing quite regularly, just because of normal
fluctuations and imprecisions in the way the metrics are calculated.

Over the last couple of weeks they've been between 0.8 and 1.35, and we
don't think we've actually had any issues with the billing system.

If they stray outside of these bounds then there's probably something
going on that needs investigating.