Move billing alerts out of "Platform Alerts"
On the user-impact Grafana dashboard.

Also renames "Tenant Alerts" to be "Other Alerts" (which is now tenant
and billing alerts, or anything with a `layer` label).

Also changes the alerting color from red to orange for the "Other
Alerts", as they're generally not as pressing.

I've already snowflaked these changes in Ireland, so you can see them
here:

https://grafana-1.cloud.service.gov.uk/d/paas-user-impact/user-impact-prod

Compare to London to see the old version:

https://grafana-1.london.cloud.service.gov.uk/d/paas-user-impact/user-impact-prod