Convert billing API perf logs to metrics
In #2029 we made paas-metrics query elasticsearch to fetch the timings
of some of the SQL queries that paas-billing's collector makes, so we
could report them to prometheus and track their performance over a
longer time frame.

This commit does the same thing, but for the queries billing API makes.
This gives us more of an insight into the user-facing performance of
billing. As the pages in paas-admin generally only make one API call,
which only makes one slow SQL query these numbers are a good proxy for
how long the pages users see take to load.

As with the other PR, because this queries logit this spans multiple
deployments (so we'll see both `prod` and `prod-lon` in production).

In my dev environment this looks like:

```
 # HELP paas_billing_api_performance_elapsed_seconds
 # TYPE paas_billing_api_performance_elapsed_seconds gauge
 paas_billing_api_performance_elapsed_seconds{deployment="leeporte",message="paas-billing.store.get-billable-event-rows-query",orgGUIDs="",rangeStart="2019-09-16",rangeStop="2019-09-17"} 0.021888704
 paas_billing_api_performance_elapsed_seconds{deployment="miki",message="paas-billing.store.get-billable-event-rows-query",orgGUIDs="",rangeStart="2019-09-16",rangeStop="2019-09-17"} 0.0217637155
 paas_billing_api_performance_elapsed_seconds{deployment="momo",message="paas-billing.store.get-billable-event-rows-query",orgGUIDs="",rangeStart="2019-09-16",rangeStop="2019-09-17"} 0.020462683
 paas_billing_api_performance_elapsed_seconds{deployment="tlwr",message="paas-billing.store.get-billable-event-rows-query",orgGUIDs="",rangeStart="2019-09-16",rangeStop="2019-09-17"} 0.019776676666666666
 paas_billing_api_performance_elapsed_seconds{deployment="towers",message="paas-billing.store.get-billable-event-rows-query",orgGUIDs="fea642e0-a0a9-47c3-ac66-e36d8ea3e92b",rangeStart="2019-09-01",rangeStop="2019-10-01"} 0.032616076
```