Increase pipeline test parallelism
The tests have been separated out for a while and are not exercising
concourse enough for us to be worried about performance. My dev env
never goes above 20% CPU for the node which runs this pipeline.

We can increase number of nodes that Ginkgo will use, which means
instead of running tests 2 at a time, it will run them 5 at time.

Additionally lower test timeouts to be closer to how long they actually
take in the custom tests.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>