Remove redundant rotation of cf-secrets
The rotate-cf-secrets.rb script has a list of secrets which it doesn't
attempt to rotate because they can't be rotated in a zero-downtime way.

As it happens, this list is the exact same as the list of all the
secrets in cf-secrets.yml (see `manifests/cf-manifest/scripts/generate-cf-secrets.rb`).
This means that none of the secrets in cf-secrets will ever get rotated,
so there's really no point in having this code.

Removing it as a first pass before we switch to using credhub to handle
these secrets.