Copy cf-secrets from S3 to credhub and use credhub
For now we're setting all the secrets in both the /concourse/main
namespace and the bosh namespace. In reality they're not all used in
both places, so we can probably split these up and simplify the code a
little.

I'm doing the credential generation directly in the pipeline instead of
hiding it away in a ruby script because I think it's simple enough that
the inline approach is clearer.

We can remove the old code and its tests because this functionality is
handled by credhub now.