transition cloud-config to paas-cf cloud-config
We need some cloud config to live in paas-bootstrap, so that we can move
the concourse manifest to be a v2 bosh manifest.

Cloud config cannot overlap, so we have to delete some things from the
cloud config in paas-cf:

- the vm definitions for compilation
- the compilation section definition

Additionally we have to specify the cpi defined in paas-bootstrap

We should name the cloud config defined in paas-cf "paas-cf" instead of
"default" otherwise it is confusing.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>