Remove cost metadata from Postgres 11 plans
This metadata field is provided to Cloud Controller by the RDS Broker,
and returned by CF API calls. It isn't displayed in `cf marketplace`.

It looks like when we made the Postgres 10 plans we copied this field
from the Postgres 9.5 plans. I'm not clear this price field was ever
correct, and we've no reason to keep it around.