Scale up routers
For 4 reasons:

- We foresee traffic increasing until the end of october, and possibly
thereafter. This has been communicated to us by
  a) the news
  b) our tenants

- There are some instances of a single Gorouter randomly throwing bad
gateways for a second.

- Gorouter currently sees daily peaks of just over 50% CPU. If we were
in the middle of rolling cells, and there were only 2 gorouter
instances, we might start dropping requests.

- If we were rolling cells and we only had 2 gorouter instances, we
would only have two AZs active

The cost impliciations of this are $210USD per environment per month as
we have not reserved more router instances

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>

Co-authored-by: Andy Paine <andy.paine@digital.cabinet-office.gov.uk>
Co-authored-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>