Allow arbitrary env names in cf_subshell_scoped_login
We now use SSO for admin accounts in all environments. Previously, it was
assumed that only the production and staging environments supported this by
default, and as such `./cf_subshell_scoped_login.sh` validated the environment
names in this way.

Now, we're safe to assume that, if the name isn't one of the known
environments, the caller was trying to log in to a development environment.