Open SSO passcode URL when using subshell login
When logging in to CloudFoundry using SSO at the command line, users are
required to provide an authentication code that they're given at theend of the
browser-based login flow. The URL to begin the flow is given at command line,
and highlighting it to open it with the browser is annoying. This commit opens
that URL in the user's browser as a convenience, before running `cf login`.