Update CFSyslogDrains alert to track adapter numbers
The alert was hardcoded to go off when any of the schedulers reached more than
250 drains. However, the documentation that we link to in the alert indicates
that the number of adapters should be:

  (cf-syslog-drain.scheduler.drains / 250) + 2

Previously, we were running only 2 instances across the 2 availability zones,
and it was running fine, so experience suggests that the below is probably fine
as a gauge of the number adapters needed:

  (sum(drains) / 250)

In this commit, I've included the
`manifests/cf-manifest/env-specific/${ENV_SPECIFIC_BOSH_VARS_FILE}` in the vars
used to generate the Prometheus manifest, so that the alerts have access to
useful variables about the scale of different components. I've then used the
`adapter_instances` variable in the alert ops file so that in future the alert
threshold will automatically scale with the number of instances of the adapter.