Fix "cf_pass is not set" on first pipeline run
The first time the `create-cloudfoundry` pipeline runs, it doesn't have
a value for `cf_pass` in credhub. This is required in the `pre-deploy`
step, which causes the pipeline to fail.

Ordinarily `cf_pass` would be generated by bosh as part of `cf-deploy`,
but since that happens later it's not much help.

Instead, this commit adds a task to generate secrets which uses credhub
to generate a value for the cf admin password (in the bosh namespace),
using the `--no-overwrite` flag to prevent this from clobbering the old
value. It then copies this secret from the bosh namespace to the
concourse namespace so it can be used in the pipeline.