Avoid CredHub issue to generate CF admin password
We added the `generate-cf-admin-password` task to generate the
CF admin password before we try to use it. We believed the
`--no-overwrite` flag would prevent `credhub generate` from changing
the value where it was already set. But in staging the secret got
changed anyway.

We've filed a CredHub issue [1] and for the time being this commit
wraps the password generation in an if statement to manually ensure
that we don't overwrite an existing secret.

[1] https://github.com/cloudfoundry-incubator/credhub/issues/60