Correctly (un)pin deployment kick off resource in Concourse
In Fly v5.6.0, the `unpin-resource` action does not take a `-v|--version`
argument, but `pin-resource` does. The way the script was written did not
distinguish between the two actions, and provided the same arguments for both.
This resulted in the `unpin-resource` action doing nothing but printing out the
version of Fly.

  $ ./bin/fly -t andyhunt unpin-resource \
    -r deployment-kick-off/deployment-timer \
    -v 2019-10-25T07:00:39.704321576+01:00

  5.6.0

This commit recognises the difference between the required arguments, so that
the `unpin-resource` actually does something.