Revert change to exchange rate
We think it's better for us to charge the exchange rate that we get when
we pay for our AWS usage, rather than the rate from the central bank.
This better reflects the costs which we're passing on to tenants.

This means we should not have changed the exchange rate to 0.78, and
going forwards we should update the rate whenever we make a new AWS
prepayment.