Specify deregistration_delay for load balancers
By default the amount of time we allow for traffic to drain to the
router instances is 300 seconds which is more than the drain_wait
property of gorouter.

We should have a deregistration_delay of less than the drain_wait (120),
e.g. 110. This was the behaviour before when we were using ELBs.

Arbitrary 60s for loggregator and doppler

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>