Vertically scale the CF scheduler VM
The scheduler VM in prod has a high load average on the active instance.

There are singleton components which run on the scheduler VM, as such
horizontally scaling it will not solve the problem.

Changing it from a medium to a large takes it from a t3.medium to an
m5.large which adds some cost, however increases the vCPU shares
allocated to it.

Additionally, when we schedule t3 instances they start with 0 CPU
credits, which means that it has no capacity to burst when under high
load. This is especially problematic when we are doing a deployment and
the scheduler VM is rolled, a new one comes up with no capacity for
bursting CPU

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>