Reduce CC API rate limit in Ireland to 20k
The previous limit in Ireland - 55k - was deliberately inflated to accommodate
a very heavy usage deployment of `paas-prometheus-exporter` that was running
the region. Since then, the exporter has been made less chatty
(https://github.com/alphagov/paas-prometheus-exporter/pull/29), causing the
deployment in question to drop to a more manageable number of calls per hour.
However, we forgot to reduce the rate limit once the fix was in place. This
commit resolves that outstanding action.

The figure of 20k/h is informed by this Kibana chart [1], which shows the top
non-admin CC api caller to be making 15k/h. I've chosen 20k here to allow some
wiggle room.

[1] https://kibana.logit.io/s/665ca355-efc3-46a2-96cf-21d31a5305bb/goto/8e10b7eb77da26e444553501aa70f56a