Increase CC API rate limit in London to 20k
The previous CC API rate limit for authenticated users in London was set at
15k/h during the initial work to set the rate limits [1]. As seen in this
Kibana chart [2], the heaviest users of the CC API in London are typically in
the region of 10k/h, well within the limits.

However, we have one user who is very frequently rate limited at 15k/h whilst
running paas-prometheus-exporter in a supported way. By increasing the rate
limit to 20k/h, we simultaneously better support this users case and bring the
two environments in to parity.

[1] https://github.com/alphagov/paas-cf/pull/1782
[2] https://kibana.logit.io/s/665ca355-efc3-46a2-96cf-21d31a5305bb/goto/8e10b7eb77da26e444553501aa70f56a