Check secrets are interpolated in pipecleaner
We want to avoid accidentally writing bits of the pipeline like this:

```
jobs:
  - name: secrets-interpolate
    plan:
      - task: bad-secrets-interpolate
        config:
          params:
            SOME_KEY_THING: some-non-interpolated-value
```

It's quite likely that if the parameter key contains `KEY` or `SECRET`
the developer meant to interpolate the value with parens, like:

```
SOME_KEY_THING: ((some-interpolated-value))
```

In some cases this could cause a broken pipeline, in others it could
cause a security issue.

It's possible that there might be some legitimate use cases that cause
this test to fail in the future - it should be fine to update the test
to ignore these specifics when they come up.