Switch billing costs from `increase` to `delta`
`increase` expects the metric to be a counter, which means it should
always either increase or reset to zero.

`paas_billing_total_costs_pounds` sounds like a counter, but it's
actually possible for it to go down (e.g. if we change our historical
exchange rate to be more favourable to tenants). In this case,
prometheus's `increase` function will assume it's a counter reset
followed by an unbelievably massive spend, which throws all the maths
out.

Switching to `delta` fixes this issue as it doesn't assume the metric is
a counter and therefore doesn't handle resets.