Apply WAF rules to app and system load balancers
We apply an AWS WAF [1] request rate rule to the system and app domain load
balancers for DDoS protection. We set the limit to allow 600,000 requests from
the same IP in a 5 minute period, based on the maximum number of requests from
any one IP address in our production environments in a 5 minute interval over
the period 2019-10-28 to 2019-11-11, rounded up to provide some leeway.

[1] https://docs.aws.amazon.com/waf/latest/APIReference/API_RateBasedRule.html