Workaround for orphaned members of groups
Our admin groups had 9 members where there wasn't a user with the
same GUID. We think this could have happened when we migrated our
admin users to log in with Google rather than a password.

Before this commit our user management tool relied on the
`/Groups/$GUID/members?returnEntities=true` endpoint. This endpoint
returns the members of a group including all the details of the
corresponding users. But this endpoint was broken because some of the
users didn't exist.

This commit works around this frustrating issue by manually expanding
members into users.