Automatically scale the Y-axis of the app instances chart
We have a chart on the user impact dashboard in Grafana which shows the number
of application instances that are running, versus how many should be running.
At the moment, the y-axis is scaled from zero to the maximum number of
instances in the time range. We have >1000 application instances in London and
>800 in Ireland, so the granularity of the y-axis is pretty coarse; 250 in
London, 200 in Ireland). This makes it difficult to see anything but the most
drastic changes in the number of instances.

Instead, in this commit we allow Grafana to pick a sensible scale for the axis
based on the data being charted.