Group members should have origin of uaa
The `origin` field a group member is not the same as the `origin`
field on a user. From https://docs.cloudfoundry.org/api/uaa/version/74.4.0/index.html#add-member:

> `origin`: The originating IDP of the entity, or "uaa" for groups and internal users

We are added users who log in with Google to these groups. Their user
has an `origin` of `google`, but their group memberships should have
an origin of `uaa`.

We've manually tested this and shown it fixes the group membership
problems we've had.