Bump the paas-audit-to-splunk app
Currently, the design of this application is not perfect. Upon crash or
restart it's job will be to re-ingest all the audit events from CF to
Splunk over and over again. This is limited by a specific date, which
now is almost a month ago, meaning the number of duplicated logs will
get insane soon...

There is an improvement, which instead of locking down to a date, will
lock down to N minutes ago, which should sanitise the ingestion but not
eliminate the problem fully.