Remove docker from credhub_shell and simplify
... and add ASCII art, of course.

People on Linux machines can't always use docker and the VPN at the same
time, because anyconnect insists on owning the entire routing table.

There's really no need for us to be using docker in this script though -
everyone's going to have SSH already installed locally, and it's not
difficult to `brew install credhub-cli`.