Make pipelines pass shellcheck 0.7.0
Shellcheck wants string variables to be quoted now, and there was a
"double quote to prevent word splitting". I've checked that the array
works the way we want by running:

```
test=(
1
2
3)
test2=(
a
b
c
"${test[@]}")
for x in ${test[*]}; do
  echo "$x"
done
```

which outputs

```
1
2
3
a
b
c
```

Which is still what we want. Shellcheck is protecting us from values in
the first array that contain whitespace.