Use paas-trusted-people user configuration file
This has a slightly different format, so:

```
- email: ''
  roles: []
```

instead of:

```
- email: ''
  cf_admin: true
```

so we're also fixing the tools/user_management scripts to work with the
new format.

paas-trusted-people is a private repo, so we have to upload it to S3 and
then download it in the pipeline (because the pipeline can't clone
private repos). The paas-trusted-people file is uploaded by a make task:

```
make upload-paas-trusted-people
```
... or just:

```
make upload-all-secrets
```