Add a script to upload trusted-people to S3
We can't clone the repo in the pipeline because it's private and we
haven't sorted out the credentials yet.

We can't store this data in credhub because it's also needed by the
create-bosh-concourse pipeline that needs to be able to run without
credhub.

This script will be duplicated in paas-bootstrap.

We've also added upload-cyber-secrets to the upload-all-secrets task -
it was missing.