Move PaaS environment config out of the Makefile
Until now we have configured PaaS CF environments directly in the
Makefile. This has worked until now, but it makes us dependent on
using Makefiles for everything.

This commit moves the environment config out of the Makefile:

  - The config is moved to a YAML file: `config/environments.yml`.

  - A script is then added which which can emit environment's env
    variables from that YAML as shell environment exports or Makefile
    variable declarations.

  - Finally the Makefile is altered to keep the same workflow but use
    this new YAML+script workflow for retrieving env variables.

The motivation for doing this now is that we plan to merge paas-cf,
paas-bootstrap and paas-release-ci. They have Make targets with the
same name and some rethinking is probably needed. Moving our config
out of the Makefiles allows us to try and find a neater solution.

This change is absolutely not meant as a very-long-term solution. The
Makefile changes are a little unpleasant, although they do work. But
this makes it possible for us to find better solutions–e.g., change
the Make targets or write a small CLI instead.