Convert gorouter {gorouter,app}_time metrics to floats
In routing-release 0.196.0, gorouter began logging two new metrics:
`gorouter_time` and `app_time`. These metrics tell operators how long a request
spent in the gorouter and the app, respectively. Previously, it was only
possible to see a `response_time` metric, which showed the time for the full
roundtrip.

We have been receiving the fields since we upgraded to routing-release 0.196.0,
but we haven't been correclty indexing them. This commit addresses that, by
converting the fields in to floats.