Scale up CF prometheus in non-dev environments
The CF Prometheus for our users should be highly available, which it
currently isn't with 1 instance. This commit makes it highly available
but only 2 instances which is probably enough

The CF Prometheus in London is currently running quite hot (85% memory)
usage due to Prometheus. This is causing it to use swap lots. In non-dev
environments make it a t3.medium

In dev, we will rarely scrape anything, so we can keep the values as
they are

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>