Use government-organisation register during create-org
The `./scripts/create-org.sh` script asks for the name of a department which
will own the CF org. Normally, we type this in by hand, but that will (in time)
lead to some inconsistency and make it difficult to search for org owned by a
department.

Instead, this commit adds a script for finding a department name based on some
input. It uses the `government-organisation` register from GOV.UK Registers to
provide the list of known departments.