Add instrumentation to ruby scripts
We don't know which Ruby scripts in our `scripts/` directory are in use. To
find out, we're adding a one-liner bit of instrumentation to log the
relative-to-paas-cf path of the script to a file every time it's run. In an
amount of time, we can look at that file to find out which scripts have been
executed in that time.