Tag `aws_elasticache_cluster_nodes_count` with the expected cache cluster id
The AWS API documentation says that CacheClusterId is the user-supplied name of
the cache cluster. However, in reality, it returns that value, plus some extra
information.

It is in the format {user-supplied}-{n}[-{m}]. Typically, our cache cluster
names have the user supplied portion with format "cf-{FNV hash of guid}", but
that's not guaranteed, so we only aim to remove the last one/two parts.

The Elasticache API returns a CacheClusterId with two numbers on the end when
the cluster is in clustered mode (yes, clusters can be in an unclustered mode),
and one when it’s not. Confusingly, Elasticache allows for a cluster in
clustered mode to have only 1 node, so we can’t rely on >1 node to mean it’s
clustered. To add more confusion, the API doesn’t indicate whether or not the
cluster is in clustered node when describing the cluster.

To get around all this, we can resort to matching the cache cluster id against
a pair of regular expressions which can tell us whether it’s got one or two
numbers on the end.