Tag individual ElastiCache cluster metrics with service GUID
ElastiCache places limits on how long the user-supplied identifier for a cache
cluster can be. The limit is shorter than a our normal format of
`{DEPLOY_ENV}-{SERVICE_GUID}`. To get around that, we hash the service GUID and
prepend `cf-`. Unfortuantely, that makes it impossible to know from the cluster
id alone which service it represents in CloudFoundry.

To help with that, we tag each ElastiCache cluster metric with the service
GUID, so that we can search our metrics for one or the other.