Label individual ElastiCache metrics with the APIs cluster id
The ElastiCache API docs says that the cluster ID is user supplied. However,
that's not quite true. It takes the user supplied bit, and tacks some info on
the end. A previous commit labelled the metrics with only the user supplied
part. This commit adds a label containing the value the API returns, so that we
can search by that too.