Expose cluster node stats at the replication group level instead
When we look in the AWS console at ElastiCache clusters, we see things at the
replication group level. In our metrics so far, we've been exposing it at the
cache cluster level. This would be confusing down the line, when HA redis
services appear >1 time in the metrics (once per cache cluster).

Instead, we now sum the node counts by replication group id.

This has the secondary benefit of making the code (and the test code) quite a
bit simpler!