Provide metadata to our services and plans
The goal is to have some nicely presented data in our PaaS Admin tool.
The ServiceBroker API supports metadata bag of keys, so we're coming up
with our custom schema for that data to be presented.

This should be non-invasive procedure provided we're not removing GUIDs
or plans.

One potentially drastic change, is me setting the `active: false` flag
to the deprecated plans, with hope that these will no longer be
available for consumption. Tell me off.