Drop problematic info-level logs from Cloud Controller job queue
We discovered that the Cloud Controller workers emit a log line starting with
"about to run job", and then a dump of the job. Unfortunately, for a couple of
job types, this results in the application manifest being logged unredacted.
When that happens, any secrets in the manifest file are exposed in plaintext
and kept in our logs for the full retention period.

This is obviously not good.

We've never spotted this before, so we can safely assume we weren't making use
of these particular log lines. To mitigate this problem, we're dropping all log
lines from any cloud controller component that begins with "about to run job".

In an ideal world, these log lines wouldn't exist at all. However, they're
coming from an upstream component so we don't have any control over that.

An issue has been raised on the project to address it.

https://github.com/cloudfoundry/cloud_controller_ng/issues/1554