Add `aws_limits_rds_instances` variable to Prometheus env-specific vars
We had a small incident where our proudction environments reached the limit on
the number of AWS RDS instances they could create in their respective regions.
This happened because we didn't have an alert for it, like we do for the number
of ElastiCache nodes and S3 buckets. We didn't have an alert because we didn't
have data.

This commit adds a new `aws_limits_rds_instances` variable, so that when we get
a metric for the number of RDS instances in a region, we can create alerts. It
also acts as a bit of documentation about our limits.