Create new RDS service types and regenerate fakes
We need a new RDS service type for wrapping communication with AWS RDS, just
like we do for other services in `paas-metrics`. This requires a new fake for
the RDS API interface, which means we also regenerate the other fakes.

In this case, their source interfaces had new methods after the dependency
update.