Alert when we reach 80% utilisation of our AWS RDS DB instances quota
On 10th February 2020 we filled our quota of AWS RDS DB instances in London.
This prevented tenants from provisioning new databases on the platform.

We filled the quota because we didn't have an alert to tell us we were close to
reaching it, and to raise a support ticket to raise it. This commit introduces
that alert, so that it doesn't happen again in future.