Use credhub shell for uploading secrets
We do not need the layers of docker to do things, we can just compose
the shell tasks

i.e. the workflow shouldnt be:

- make dev upload-all-secrets => do all the things

instead we can rely on people to compose the steps correctly, with
appropriate pointers along the way. This way the code is easier to
follow and is less fragile to change:

- make dev credhub
> credhub login --sso
> make upload-all-secrets

(where > represents a subshell)

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>