Refactor Elasticache broker operations file
We're going to be adding plans for Redis 4.x and 5.x, so the file is going to
get quite a bit longer.

Right now, the configuration of each plan is repeated, making the file
unnecessarily long. This commit extracts the repeated parts of config in to
YAML anchors, and uses them to DRY up the config and make it more readable.