Set `maximum_app_disk_in_mb` in `cloud_controller_worker`
Previously, we had set a maximum amount of disk space an app instance could
request to 4gb, using the property in the `cloud_controller_ng` job. However,
as this GitHub issue [1] outlines, the value needs to be set on both
`cloud_controller_ng` and `cloud_controller_worker` in order to support
synchronous AND asynchronous app updates (n.b. the vast majority of updates are
async)

[1] https://github.com/cloudfoundry/cloud_controller_ng/issues/1405