Set Cloud Controller Worker job timeouts
The default timeout for a job in the cloud controller worker queue is 4 hours.
This is a very long time for a job to be hanging about when it isn't
completing. This commit sets the timeout, and that of deleting from the
blobstore, to 1 hour. This should provide time for a long running job to
complete, without being so long as to clutter the queue with jobs that will
never complete.